+++ ChatGP-Terminal/Src/libGPT/libGPT.c
@@ -40,7 +40,7 @@ typedef struct Messages{
 static Messages *historyContext=NULL;
 static int contHistoryContext=0;
 static int maxHistoryContext=0;
-long int recvTimeOut=SOCKET_RECV_TIMEOUT_MS;
+static long int recvTimeOut=SOCKET_RECV_TIMEOUT_MS;
 
 int libGPT_set_timeout(long int timeOut){
 	if(timeOut<0) return LIBGPT_RECV_TIMEOUT_ERROR;