+++ ChatGP-Terminal/Releases/CHANGELOG.md
@@ -6,6 +6,8 @@
 Features:
 - Added option '--top-p'. It allows to incorporate a 'Top P' value. Default 1.0.
 - Added option 'tp; [value]' when prompting, for changing the 'top p' parameter.
+- Added option '--timeout'. For setting the timeout of responses (in ms). Default 60000.
+- Added option 'to; [value]' when prompting, for changing the timeout.
 - //TODO _(under evaluation)_ option '--alert-finished-status' for showing the finished status only if finished status != "Stop".
 
 Improvements: